FV3Core Changes for GTC (#584)
This PR implements a number of changes that allow fv3core to validate with the GTC compiler toolchain of gt4py.

- Code changes:
* Add explicit in/out versions of delp, pt, and w fields in c_sw.fill_corners_delp_pt_w
* Add new temporaries uf0 and vf0 formerly reassigned to uf and vf in c_sw.divergence_corner
* Split last compute block in d_sw.heat_source_from_vorticity_damping to d_sw.update_u_and_v
* Add explicit in/out versions of q field to del2cubed.corner_fill
* Alternate iteration order between PARALLEL and FORWARD in del2cubed.corner_fill to prevent incorrect merging.
* Inline self._stretched_grid at compile time in divergence_damping
* Remove unnecessary temporaries in map_single.lagrangian_contributions stencil
* Inline self._hydrostatic as compile time if in ray_fast
* Update fill_corners calls in updatedzc
* Replace reads with horizontal offsets within regions in xppm.compute_x_flux and yppm.compute_y_flux with temporaries
* Add explicit q_in and q_corner fields to corners.fill_corners_2cells_x_stencil and corners.fill_corners_2cells_y_stencil
* Add explicit in/out versions of x and y fields in corners.fill_corners_dgrid_defn

- Infrastructure changes:
* Change axis offset endpoints from np.iinfo(np.int32).min/max to np.iinfo(np.int16).min/max

- Requirements changes:
* Update GT4PY_VERSION to v35

- Minor changes:
* Remove redundant logging initialization in gt4py_utils and rename from fv3ser to fv3core
* Always call output_data = gt_utils.asarray(output[varname]) in test_translate methods to ensure cupy storages are always copied to numpy storages before comparing results

+++ fv3core/utils/grid.py
@@ -1141,28 +1141,28 @@ def _old_grid_axis_offsets(
         origin_offset = grid.is_ - origin[0]
         i_start = gtscript.I[0] + proc_offset + origin_offset
     else:
-        i_start = gtscript.I[0] - np.iinfo(np.int32).max
+        i_start = gtscript.I[0] - np.iinfo(np.int16).max
 
     if grid.east_edge:
         proc_offset = grid.npx + grid.halo - 2 - grid.global_is
         endpt_offset = (grid.is_ - origin[0]) - domain[0] + 1
         i_end = gtscript.I[-1] + proc_offset + endpt_offset
     else:
-        i_end = gtscript.I[-1] + np.iinfo(np.int32).max
+        i_end = gtscript.I[-1] + np.iinfo(np.int16).max
 
     if grid.south_edge:
         proc_offset = grid.js - grid.global_js
         origin_offset = grid.js - origin[1]
         j_start = gtscript.J[0] + proc_offset + origin_offset
     else:
-        j_start = gtscript.J[0] - np.iinfo(np.int32).max
+        j_start = gtscript.J[0] - np.iinfo(np.int16).max
 
     if grid.north_edge:
         proc_offset = grid.npy + grid.halo - 2 - grid.global_js
         endpt_offset = (grid.js - origin[1]) - domain[1] + 1
         j_end = gtscript.J[-1] + proc_offset + endpt_offset
     else:
-        j_end = gtscript.J[-1] + np.iinfo(np.int32).max
+        j_end = gtscript.J[-1] + np.iinfo(np.int16).max
 
     return {
         "i_start": i_start,
@@ -1185,26 +1185,26 @@ def _grid_indexing_axis_offsets(
     if grid.west_edge:
         i_start = gtscript.I[0] + grid.origin[0] - origin[0]
     else:
-        i_start = gtscript.I[0] - np.iinfo(np.int32).max
+        i_start = gtscript.I[0] - np.iinfo(np.int16).max
 
     if grid.east_edge:
         i_end = (
             gtscript.I[-1] + (grid.origin[0] + grid.domain[0]) - (origin[0] + domain[0])
         )
     else:
-        i_end = gtscript.I[-1] + np.iinfo(np.int32).max
+        i_end = gtscript.I[-1] + np.iinfo(np.int16).max
 
     if grid.south_edge:
         j_start = gtscript.J[0] + grid.origin[1] - origin[1]
     else:
-        j_start = gtscript.J[0] - np.iinfo(np.int32).max
+        j_start = gtscript.J[0] - np.iinfo(np.int16).max
 
     if grid.north_edge:
         j_end = (
             gtscript.J[-1] + (grid.origin[1] + grid.domain[1]) - (origin[1] + domain[1])
         )
     else:
-        j_end = gtscript.J[-1] + np.iinfo(np.int32).max
+        j_end = gtscript.J[-1] + np.iinfo(np.int16).max
 
     return {
         "i_start": i_start,
