+++ tests/main/test_grid.py
@@ -86,19 +86,19 @@ def test_axis_offsets(
     if west_edge:
         assert axis_offsets["i_start"] == i_start
     else:
-        assert axis_offsets["i_start"] == gtscript.I[0] - np.iinfo(np.int32).max
+        assert axis_offsets["i_start"] == gtscript.I[0] - np.iinfo(np.int16).max
     if east_edge:
         assert axis_offsets["i_end"] == i_end
     else:
-        assert axis_offsets["i_end"] == gtscript.I[-1] + np.iinfo(np.int32).max
+        assert axis_offsets["i_end"] == gtscript.I[-1] + np.iinfo(np.int16).max
     if south_edge:
         assert axis_offsets["j_start"] == j_start
     else:
-        assert axis_offsets["j_start"] == gtscript.J[0] - np.iinfo(np.int32).max
+        assert axis_offsets["j_start"] == gtscript.J[0] - np.iinfo(np.int16).max
     if north_edge:
         assert axis_offsets["j_end"] == j_end
     else:
-        assert axis_offsets["j_end"] == gtscript.J[-1] + np.iinfo(np.int32).max
+        assert axis_offsets["j_end"] == gtscript.J[-1] + np.iinfo(np.int16).max
 
 
 @pytest.mark.parametrize(
